+++ ReleaseNotes.md
@@ -15,7 +15,7 @@ Here come the actual release notes:
 upcoming
 
 ## Features
-
+ * ME-46 Support Ubuntu 16.04 server with metasfresh server installer
  * #369 request report
  * #361 Request change for customer service
  * #377 Implement executed SQLs tracing
@@ -33,6 +33,7 @@ upcoming
  * #7 provide "grid-align" for layout elements
 
 ## Fixes
+ * #380 duplicate lines in inout
  * #348 Sort tabs shall consider Link column and parent link column if set
  * #330 Process's RefreshAllAfterExecution does not work when the record was moved
  * #327 Got NPE when completing a drafted order